api_challenge/cake_creation/urls.py
from rest_framework.routers import SimpleRouter
from cake_creation.views import CakeView
router = SimpleRouter(trailing_slash=False)

router.register(r"cake_system", CakeView)

urlpatterns = []
urlpatterns+=router.urls
