api_challenge/requirements.txt
asgiref==3.7.2
Django==4.2.7
django-filter==23.3
djangorestframework==3.14.0
drf-yasg==1.21.7
exceptiongroup==1.1.3
factory-boy==3.3.0
Faker==20.0.3
inflection==0.5.1
iniconfig==2.0.0
numpy==1.26.2
packaging==23.2
pluggy==1.3.0
pytest==7.4.3
pytest-django==4.7.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
